chore: remove functional import
